RUST DISEASE DETECTION SYSTEM/requirements.txt
# Core Dependencies
tensorflow>=2.5.0
numpy>=1.19.2
Pillow>=8.3.1
matplotlib>=3.4.3
opencv-python>=4.5.3
scikit-image>=0.18.3
scikit-learn>=0.24.2
seaborn>=0.11.2

# Image Processing
tifffile>=2021.8.30

# Deep Learning
tensorflow-gpu>=2.5.0  # If using GPU
keras>=2.4.3

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2

# Image Processing Libraries
opencv-python>=4.5.3
scikit-image>=0.18.3

# Machine Learning Tools
scikit-learn>=0.24.2

# Optional Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional Dependencies
h5py>=3.1.0
pandas>=1.3.0